Change Main.js to Blockchain.js
