+++ main.js
@@ -11,13 +11,3 @@
  require('./app/components/blockchain');
  require('./app/components/block');
 
- import BlockChain from './app/components/blockchain';
-
-//let BlockChain2 = BlockChain();
-let Block = new Block();
-//var Block2 = new Block("12/15/2018", {amount:20});
-BlockChain.addBlock(Block);
-//jkChainObjext.addBlock(new blockObject.block("12/16/2018",{amount:67}));
-
-console.log(JSON.stringify(BlockChain,null,4));
-console.log("Is blockchain valid?" + BlockChain.checkValid());
